+++ R/wikidata.R
@@ -9,6 +9,7 @@
 #' @param wait A numeric value passed into `Sys.sleep()` to slow down sequence
 #'   of requests (and avoid denial of service). Defaults to 100.
 #' @param progress Whether to show progress bar (`logical` value).
+#' @export
 #' @examples
 #' \donttest{
 #' dbpedia_ids <- c(
@@ -79,6 +80,79 @@ dbpedia_get_wikidata_uris <- function(x, optional, endpoint, limit = 100, wait =
       retval_li[[i]][["wikidata_uri"]]
     )
     colnames(retval_li[[i]])[1] <- "dbpedia_uri"
+  }
+  
+  if (progress) cli_progress_done()
+  
+  do.call(rbind, retval_li)
+}
+
+
+#' Query Wikidata endpoint for additional information.
+#' 
+#' This is a wrapper for `WikidataQueryServiceR::query_wikidata()` to get
+#' additional information for known wikidata IDs.
+#' 
+#' @return A `tibble`.
+#' @param x A vector of wikidata ids.
+#' @param id Wikidata ID for information to retrieve (`character` vector).
+#' @param limit Maximum number of wikidata IDs to be sent to endpoint at a time.
+#' @param progress Whether to show progress information (`logical` value).
+#' @param wait A numeric value - slow down requests to avoid denial of service.
+#' @export
+#' @examples
+#' \donttest{
+#' wikidata_ids <- c("Q1741365", "Q3840", "Q437")
+#' wikidata_resolve_dbpedia_uri(
+#'   wikidata_ids,
+#'   id = "P439", # German municipality key
+#'   wait = 0,
+#'   limit = 2,
+#'   progress = TRUE
+#' )
+#' }
+wikidata_query <- function(x, id, limit = 100L, wait = 1, progress = FALSE){
+  
+  if (!requireNamespace("WikidataQueryServiceR", quietly = TRUE)){
+    stop("R package WikidataQueryServiceR required but not available. ")
+  }
+  
+  stopifnot(
+    is.vector(x), is.character(x),
+    is.character(id), length(id) == 1L,
+    is.numeric(limit), limit > 0,
+    is.numeric(wait), wait > 0, length(wait) == 1L,
+    is.logical(progress), length(progress) == 1L
+  )
+  
+  template <- 'SELECT ?item ?label ?key ?keyLabel
+        WHERE {
+        VALUES ?item { %s }
+        OPTIONAL { ?item wdt:%s ?key . }
+        ?item rdfs:label ?label
+          filter(lang(?label) = "de")
+        SERVICE wikibase:label { bd:serviceParam wikibase:language "de". }
+      }'
+  
+  chunks <- as_chunks(x = x, size = limit)
+  retval_li <- list()
+  
+  if (progress) cli_progress_bar("Tasks", total = length(chunks), type = "tasks")
+  for (i in 1L:length(chunks)){
+    cli_progress_update()
+    query <- sprintf(
+      template,
+      paste0("wd:", chunks[[i]], collapse = " "),
+      id
+    )
+    
+    Sys.sleep(wait)
+    
+    retval_li[[i]] <- WikidataQueryServiceR::query_wikidata(
+      sparql_query = query,
+      format = "simple"
+    )
+    colnames(retval_li[[i]])[1] <- "wikidata_id"
   }
   
   if (progress) cli_progress_done()